+++ tests/test_lib.py
@@ -4,30 +4,26 @@ import re
 from collections import Counter, defaultdict, namedtuple
 from functools import partial
 from itertools import groupby
+from html import unescape
 
 import pytest
-from rich.columns import Columns
-from rich.traceback import install
-from rich_tables.utils import border_panel, make_console, make_difftext, new_table, wrap
-
 from beetsplug.bandcamp import BandcampPlugin
 from beetsplug.bandcamp._metaguru import Metaguru
+from rich.columns import Columns
+from rich.traceback import install
+
+from rich_tables.utils import border_panel, make_console, make_difftext, new_table, wrap
 
 pytestmark = pytest.mark.lib
 
 target_dir = "dev"
-compare_against = "cur"
+compare_against = "f9582f1"
 if not os.path.exists(target_dir):
     os.makedirs(target_dir)
 install(show_locals=True, extra_lines=8, width=int(os.environ.get("COLUMNS", 150)))
-console = make_console()
+console = make_console(stderr=True)
 
-testfiles = os.listdir("tests/jsons")
-
-
-@pytest.fixture(scope="session")
-def config(request):
-    return BandcampPlugin().config.flatten()
+testfiles = list(filter(lambda x: x.endswith("json"), os.listdir("jsons")))
 
 
 @pytest.fixture(params=testfiles)
@@ -36,8 +32,50 @@ def file(request):
 
 
 Oldnew = namedtuple("Oldnew", ["old", "new", "diff"])
+
+
 oldnew = defaultdict(list)
+
+
+@pytest.fixture(scope="session")
+def _report():
+    yield
+    cols = []
+    for field in set(oldnew.keys()) - {"comments", "genre"}:
+        field_diffs = sorted(oldnew[field], key=lambda x: x.new)
+        if not field_diffs:
+            continue
+        tab = new_table()
+        for new, all_old in groupby(field_diffs, lambda x: x.new):
+            tab.add_row(
+                " | ".join(
+                    map(
+                        lambda x: (f"{x[1]} x " if x[1] > 1 else "")
+                        + wrap(x[0], "b s red"),
+                        Counter(map(lambda x: x.old, all_old)).items(),
+                    )
+                ),
+                wrap(new, "b green"),
+            )
+        cols.append(border_panel(tab, title=field))
+
+    console.print("")
+    console.print(Columns(cols, expand=True))
+
+    stats_table = new_table("field", "#", border_style="white")
+    for field, count in sorted(stats_map.items(), key=lambda x: x[1], reverse=True):
+        stats_table.add_row(field, str(count))
+    if stats_table.rows:
+        stats_table.add_row("total", str(len(testfiles)))
+        console.print(stats_table)
+
+
 stats_map = defaultdict(lambda: 0)
+
+
+@pytest.fixture(scope="module")
+def config(request):
+    yield BandcampPlugin().config.flatten()
 
 
 def do_key(table, key: str, before, after) -> None:
@@ -48,11 +86,11 @@ def do_key(table, key: str, before, after) -> None:
         difftext = ""
         stats_map[key] += 1
         if key == "genre":
-            before, after = set(before.split(", ")), set(after.split(", "))
+            before_set, after_set = set(before.split(", ")), set(after.split(", "))
             common, gone, added_new = (
-                before & after,
-                before - after,
-                after - before,
+                before_set & after_set,
+                before_set - after_set,
+                after_set - before_set,
             )
             diffparts = list(map(partial(wrap, tag="b #111111"), sorted(common)))
             if gone:
@@ -88,13 +126,17 @@ def compare(old, new) -> bool:
             do_key(table, key, str(old.get(key, "")), str(new.get(key, "")))
 
     if table.rows:
+        console.print("")
         console.print(border_panel(table, title=title, subtitle=album_id))
         pytest.fail(pytrace=False)
 
 
+@pytest.mark.usefixtures("_report")
 def test_file(file, config):
-    meta = json.load(open(os.path.join("tests/jsons", file)))
-    guru = Metaguru(meta, config)
+    meta_file = os.path.join("jsons", file)
+    tracks_file = os.path.join("jsons", file.replace(".json", ".tracks"))
+    meta = open(meta_file).read() + "\n" + unescape(unescape(open(tracks_file).read()))
+    guru = Metaguru.from_html(meta, config)
 
     if "_track_" in file:
         new = guru.singleton
@@ -106,53 +148,3 @@ def test_file(file, config):
 
     old = json.load(open(os.path.join(compare_against, file)))
     compare(old, new)
-
-
-def test_all(config):
-    for testfile in sorted(testfiles):
-        meta = json.load(open(os.path.join("tests/jsons", testfile)))
-        guru = Metaguru(meta, config)
-        if "_track_" in testfile:
-            new = guru.singleton
-        else:
-            new = guru.album
-
-        target = os.path.join(target_dir, testfile)
-        json.dump(new, open(target, "w"), indent=2)
-        try:
-            old = json.load(open(os.path.join(compare_against, testfile)))
-        except FileNotFoundError:
-            old = {}
-        try:
-            compare(old, new)
-        except:  # noqa
-            pass
-
-    cols = []
-    for field in set(oldnew.keys()) - {"comments"}:
-        field_diffs = sorted(oldnew[field], key=lambda x: x.new)
-        if not field_diffs:
-            continue
-        tab = new_table()
-        for new, all_old in groupby(field_diffs, lambda x: x.new):
-            tab.add_row(
-                " | ".join(
-                    map(
-                        lambda x: (f"{x[1]} x " if x[1] > 1 else "")
-                        + wrap(x[0], "b s red"),
-                        Counter(map(lambda x: x.old, all_old)).items(),
-                    )
-                ),
-                wrap(new, "b green"),
-            )
-        cols.append(border_panel(tab, title=field))
-
-    console.print(Columns(cols, expand=True))
-
-    stats_table = new_table("field", "#", border_style="white")
-    for field, count in sorted(stats_map.items(), key=lambda x: x[1], reverse=True):
-        stats_table.add_row(field, str(count))
-    if stats_table.rows:
-        stats_table.add_row("total", str(len(testfiles)))
-        console.print(stats_table)
-        pytest.fail()